refactor(sound): lift the GF1 out of the kernel into //lib:sound
The GUS chip is now a passive state machine with no idea what a vPIC, an
8237 or a guest address is: //lib:sound's Gf1 takes its clock as an
argument, reports that it wants service instead of raising, and is handed
DMA bytes the host already fetched. What stayed in kernel vgus.rs is
exactly what could not go: ULTRASND presence/wiring, the DMA cycle, the
interrupt line, the clock, and the trace ring's mutable statics.

//lib:sampler becomes //lib:sound (engine/ + gus.rs); design in
SOUND_LIB_DESIGN.md.
