+++ kernel/BUILD.bazel
@@ -53,7 +53,7 @@ rust_library(
     deps = [
         "//arch-abi",
         "//lib",
-        "//lib:sampler",
+        "//lib:sound",
         "//third_party/lwext4:lwext4",
         "//third_party/lwext4:lwext4-sys",
         # no_std-patched http_archive (not @crates): one flavour serves both
@@ -179,7 +179,7 @@ rust_binary(
         ":kernel",
         "//arch-interp",
         "//lib",
-        "//lib:sampler",
+        "//lib:sound",
         "//third_party/lwext4:lwext4",
         "//third_party/lwext4:lwext4-sys",
         "@lwext4//:lwext4_host",
@@ -201,7 +201,7 @@ rust_binary(
         ":kernel",
         "//arch-interp:arch-interp-kvm",
         "//lib",
-        "//lib:sampler",
+        "//lib:sound",
         "//third_party/lwext4:lwext4",
         "//third_party/lwext4:lwext4-sys",
         "@lwext4//:lwext4_host",
@@ -273,7 +273,7 @@ rust_binary(
         ":kernel",
         "//arch-interp",
         "//lib",
-        "//lib:sampler",
+        "//lib:sound",
         "//third_party/lwext4:lwext4",
         "//third_party/lwext4:lwext4-sys",
         "@lwext4//:lwext4_host",
